Added 'historic' record for each city
TemperatureSummary now has two optional properties that helps to store the Historic average.

Each city now have an 'historic' key in the dictionary that it's also a TemperatureSummary object, and stores the historical low, high and average temperature of the city.

The average historical temperature its calculated by the division between the accumAvr and regCount.